Moved FIles around, changed naming, and printing
Used reflections to order runs by year and day, renamed files and packages
