+++ AOCFramework/src/main/java/aoc/framework/DaySolution.java
@@ -1,5 +1,11 @@
 package aoc.framework;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
 public @interface DaySolution {
+    int year();
+
     int day();
 }
