+++ AOCSolutions/src/main/resources/202301/input.txt
