+++ AOCSolutions/src/main/resources/202301/input_short_part_1.txt
