+++ AOCSolutions/src/main/resources/202302/input.txt
