+++ AOCSolutions/src/main/resources/202302/input_short_part_1.txt
