+++ AOCSolutions/src/main/resources/202303/input.txt
