+++ AOCSolutions/src/main/resources/202303/input_short_part_2.txt
