--- AOCSolutions/src/main/resources/day01/input_short_part_1.txt
@@ -1,4 +1,0 @@
-1abc2
-pqr3stu8vwx
-a1b2c3d4e5f
-treb7uchet
