--- AOCSolutions/src/main/resources/day03/input_short_part_2.txt
@@ -1,10 +1,0 @@
-467..114..
-...*......
-..35..633.
-......#...
-617*......
-.....+.58.
-..592.....
-......755.
-...$.*....
-.664.598..
